+++ docs/conf.py
@@ -12,7 +12,7 @@ sys.path.insert(0, os.path.abspath('../src'))
 project = 'LayeredCompModel'
 copyright = '2026, John Kossa'
 author = 'John Kossa'
-release = '0.1.0'
+release = '0.3.0'
 
 # -- General configuration ---------------------------------------------------
 extensions = [